Code cleanup fixes
- Moved `persistent` module to root
- Renamed `Ref` to `SafeRef`
- Renamed `WeakRef` to `SafeWeakRef`
- Renamed `rf` module to `js_ref`
- Moved `*/mod.rs` to `*.rs`
- Moved `RustAllocator` to submodule
- Moved `FileResolver` to submodule
- Moved `ScriptLoader` to submodule
- Moved `NativeLoader` to submodule
