+++ src/loader.rs
@@ -1,6 +1,17 @@
 use crate::{qjs, BeforeInit, Ctx, Error, Module, Result};
-use relative_path::{RelativePath, RelativePathBuf};
+use relative_path::RelativePath;
 use std::{ffi::CStr, ptr};
+
+mod file_resolver;
+pub use file_resolver::FileResolver;
+
+mod script_loader;
+pub use script_loader::ScriptLoader;
+
+#[cfg(feature = "dyn-load")]
+mod native_loader;
+#[cfg(feature = "dyn-load")]
+pub use native_loader::NativeLoader;
 
 /// Module resolver trait
 ///
@@ -149,245 +160,6 @@ impl LoaderHolder {
             ptr::null_mut()
         })
     }
-}
-
-/// The file module resolver
-///
-/// This resolver can be used as the nested backing resolver in user-defined resolvers.
-#[derive(Debug)]
-pub struct FileResolver {
-    paths: Vec<RelativePathBuf>,
-    patterns: Vec<String>,
-}
-
-impl FileResolver {
-    /// Add search path for modules
-    pub fn add_path<P: Into<RelativePathBuf>>(&mut self, path: P) -> &mut Self {
-        self.paths.push(path.into());
-        self
-    }
-
-    /// Add module file pattern
-    pub fn add_pattern<P: Into<String>>(&mut self, pattern: P) -> &mut Self {
-        self.patterns.push(pattern.into());
-        self
-    }
-
-    /// Add support for native modules
-    pub fn add_native(&mut self) -> &mut Self {
-        #[cfg(target_family = "windows")]
-        self.add_pattern("{}.dll");
-
-        #[cfg(target_vendor = "apple")]
-        self.add_pattern("{}.dylib").add_pattern("lib{}.dylib");
-
-        #[cfg(all(target_family = "unix"))]
-        self.add_pattern("{}.so").add_pattern("lib{}.so");
-
-        self
-    }
-
-    /// Build resolver
-    pub fn build(&self) -> Self {
-        Self {
-            paths: self.paths.clone(),
-            patterns: self.patterns.clone(),
-        }
-    }
-
-    fn try_patterns(&self, path: &RelativePath) -> Option<RelativePathBuf> {
-        if let Some(extension) = &path.extension() {
-            if !is_file(path) {
-                return None;
-            }
-            // check for known extensions
-            self.patterns
-                .iter()
-                .find(|pattern| {
-                    let path = RelativePath::new(pattern);
-                    if let Some(known_extension) = &path.extension() {
-                        known_extension == extension
-                    } else {
-                        false
-                    }
-                })
-                .map(|_| path.to_relative_path_buf())
-        } else {
-            // try with known patterns
-            self.patterns
-                .iter()
-                .filter_map(|pattern| {
-                    let name = pattern.replace("{}", path.file_name()?);
-                    let file = path.with_file_name(&name);
-                    if is_file(&file) {
-                        Some(file)
-                    } else {
-                        None
-                    }
-                })
-                .next()
-        }
-    }
-}
-
-impl Default for FileResolver {
-    fn default() -> Self {
-        Self {
-            paths: vec![],
-            patterns: vec!["{}.js".into()],
-        }
-    }
-}
-
-impl Resolver for FileResolver {
-    fn resolve<'js>(&mut self, _ctx: Ctx<'js>, base: &str, name: &str) -> Result<String> {
-        let path = if !name.starts_with('.') {
-            self.paths
-                .iter()
-                .filter_map(|path| {
-                    let path = path.join_normalized(name);
-                    self.try_patterns(&path)
-                })
-                .next()
-        } else {
-            let path = RelativePath::new(base);
-            let path = if let Some(dir) = path.parent() {
-                dir.join_normalized(name)
-            } else {
-                name.into()
-            };
-            self.try_patterns(&path)
-        }
-        .ok_or_else(|| Error::resolving::<_, _, &str>(base, name, None))?;
-
-        Ok(path.to_string())
-    }
-}
-
-/// The script module loader
-///
-/// This loader can be used as the nested backing loader in user-defined loaders.
-#[derive(Debug)]
-pub struct ScriptLoader {
-    extensions: Vec<String>,
-}
-
-impl ScriptLoader {
-    /// Add file extensions
-    pub fn add_extension<X: Into<String>>(&mut self, extension: X) -> &mut Self {
-        self.extensions.push(extension.into());
-        self
-    }
-
-    /// Build loader
-    pub fn build(&self) -> Self {
-        Self {
-            extensions: self.extensions.clone(),
-        }
-    }
-}
-
-impl Default for ScriptLoader {
-    fn default() -> Self {
-        Self {
-            extensions: vec!["js".into()],
-        }
-    }
-}
-
-impl Loader for ScriptLoader {
-    fn load<'js>(&mut self, ctx: Ctx<'js>, path: &str) -> Result<Module<'js, BeforeInit>> {
-        if !check_extensions(&path, &self.extensions) {
-            return Err(Error::loading::<_, &str>(path, None));
-        }
-
-        let source: Vec<_> = std::fs::read(&path)?;
-        ctx.compile_only(path, source)
-    }
-}
-
-#[cfg(feature = "dyn-load")]
-/// The native module loader
-///
-/// This loader can be used as the nested backing loader in user-defined loaders.
-///
-/// # Features
-/// This struct is only available if the `dyn-load` features is enabled.
-#[derive(Debug)]
-pub struct NativeLoader {
-    extensions: Vec<String>,
-    libs: Vec<dlopen::raw::Library>,
-}
-
-#[cfg(feature = "dyn-load")]
-impl NativeLoader {
-    /// Add file extension
-    pub fn add_extension<X: Into<String>>(&mut self, extension: X) -> &mut Self {
-        self.extensions.push(extension.into());
-        self
-    }
-
-    /// Build loader
-    pub fn build(&self) -> Self {
-        Self {
-            extensions: self.extensions.clone(),
-            libs: Vec::new(),
-        }
-    }
-}
-
-#[cfg(feature = "dyn-load")]
-impl Default for NativeLoader {
-    fn default() -> Self {
-        let mut loader = Self {
-            extensions: Vec::new(),
-            libs: Vec::new(),
-        };
-
-        #[cfg(target_family = "windows")]
-        loader.add_extension("dll");
-
-        #[cfg(all(target_family = "unix"))]
-        loader.add_extension("so");
-
-        #[cfg(target_vendor = "apple")]
-        loader.add_extension("dylib");
-
-        loader
-    }
-}
-
-#[cfg(feature = "dyn-load")]
-impl Loader for NativeLoader {
-    fn load<'js>(&mut self, ctx: Ctx<'js>, path: &str) -> Result<Module<'js, BeforeInit>> {
-        use dlopen::raw::Library;
-        use std::ffi::CString;
-
-        if !check_extensions(&path, &self.extensions) {
-            return Err(Error::loading::<_, &str>(path, None));
-        }
-
-        type LoadFn =
-            unsafe extern "C" fn(*mut qjs::JSContext, *const qjs::c_char) -> *mut qjs::JSModuleDef;
-
-        let lib = Library::open(&path).map_err(|_| Error::Unknown)?;
-        let load_fn: LoadFn =
-            unsafe { lib.symbol("js_init_module") }.map_err(|_| Error::Unknown)?;
-
-        let name = CString::new(path)?;
-        let ptr = unsafe { load_fn(ctx.ctx, name.as_ptr()) };
-
-        if ptr.is_null() {
-            Err(Error::Unknown)
-        } else {
-            self.libs.push(lib);
-            Ok(unsafe { Module::from_module_def(ctx, ptr) })
-        }
-    }
-}
-
-fn is_file<P: AsRef<RelativePath>>(path: P) -> bool {
-    path.as_ref().to_path(".").is_file()
 }
 
 fn check_extensions(name: &str, extensions: &[String]) -> bool {
