+++ src/property.rs
@@ -239,7 +239,7 @@ mod test {
     #[test]
     fn property_with_getter_and_setter() {
         test_with(|ctx| {
-            let val = safe_ref::Ref::new(StdString::new());
+            let val = SafeRef::new(StdString::new());
             let obj = Object::new(ctx).unwrap();
             obj.prop(
                 "key",
