+++ src/value/atom.rs
@@ -1,8 +1,4 @@
-use crate::{
-    qjs,
-    value::{self, rf::JsStringRef, String, Value},
-    Ctx, Error, Result,
-};
+use crate::{qjs, value, Ctx, Error, JsStringRef, Result, String, Value};
 use std::{ffi::CStr, mem, string::String as StdString};
 
 /// An atom is value representing the name of a variable of an objects and can be created
