+++ src/value/function/ffi.rs
@@ -1,5 +1,5 @@
 use super::ArgsValue;
-use crate::{context::Ctx, qjs, value::handle_panic, Result, Value};
+use crate::{qjs, value::handle_panic, Ctx, Result, Value};
 use std::{panic, panic::AssertUnwindSafe, ptr};
 
 static mut FUNC_CLASS: qjs::JSClassID = 0;
