+++ src/value/object.rs
@@ -1,8 +1,6 @@
 use crate::{
-    qjs,
-    value::{self, rf::JsObjectRef},
-    Array, Atom, Ctx, Error, FromAtom, FromIteratorJs, FromJs, Function, IntoAtom, IntoJs, Result,
-    Value,
+    qjs, value, Array, Atom, Ctx, Error, FromAtom, FromIteratorJs, FromJs, Function, IntoAtom,
+    IntoJs, JsObjectRef, Result, Value,
 };
 use std::{
     iter::{IntoIterator, Iterator},
