+++ src/value/string.rs
@@ -1,4 +1,4 @@
-use crate::{context::Ctx, qjs, value, value::rf::JsStringRef, Error, Result, StdString, Value};
+use crate::{qjs, value, Ctx, Error, JsStringRef, Result, StdString, Value};
 use std::{mem, slice, str};
 
 /// Rust representation of a javascript string.
